refactor(api): move usecase from lib to src
