+++ api/src/identity-access-management/domain/usecases/update-expired-password.usecase.js
@@ -2,9 +2,9 @@ import lodash from 'lodash';
 
 const { get } = lodash;
 
-import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../src/identity-access-management/domain/constants/identity-providers.js';
-import { ForbiddenAccess, UserNotFoundError } from '../../../src/shared/domain/errors.js';
-import { logger } from '../../../src/shared/infrastructure/utils/logger.js';
+import { ForbiddenAccess, UserNotFoundError } from '../../../shared/domain/errors.js';
+import { logger } from '../../../shared/infrastructure/utils/logger.js';
+import { NON_OIDC_IDENTITY_PROVIDERS } from '../constants/identity-providers.js';
 
 const updateExpiredPassword = async function ({
   passwordResetToken,
